union de todo el back end

+++ back/controllers/postControllers.js
@@ -4,12 +4,12 @@ const {commentModel,postModel,userModel} = require('../models/associations');
 const { validationResult } = require("express-validator");
 //Require fs module
 const fs = require('fs');
-
+const db = require('../database/db');
 //get all posts
 const getAllPosts = async(req,res) => {
     try {
         const posts = await postModel.findAll({
-            include:[{model:userModel,attributes:["id","user","avatar"]},{model:commentModel,attributes:["id","content","commentImg","cantReactions"]}],
+            include:[{model:userModel,attributes:["id","user","avatar"]},{model:commentModel,attributes:["id","content","commentImg","cantReactions"],include:{model:userModel,attributes:["id",'user','avatar']}}],
             attributes:["id","content","postImg","cantReactions"]
         })
         res.json(posts)
@@ -39,12 +39,15 @@ const createPost = async(req,res) => {
     console.log(req.files);
     if(!errors.isEmpty()){
         fs.unlinkSync(`public/posts/${req.files[0].filename}`)
-        return res.json(errors.mapped())
+        res.json(errors.mapped())
     }else{
        try {
+        await db.query("ALTER TABLE posts AUTO_INCREMENT = 1");
         await postModel.create({
             content : content,
-            postImg : postImg
+            postImg : postImg,
+            //where esta userId
+            usersId: req.session.userLog.id
         })
         res.json({
             'message' : 'Post successfully created'
@@ -59,11 +62,16 @@ const createPost = async(req,res) => {
 const updatePost = async (req,res) =>{
     let errors = validationResult(req);
     console.log(errors);
+    //creo esto porque si no hay una imagen al editar que se quede con la antigua imagen del post
+    const postById = await postModel.findByPk(req.params.id)
     let {content} = req.body;
-    let postImg = req.files[0] ? req.files[0].filename : 'default.png' 
+    let postImg = req.files[0] ? req.files[0].filename : postById.avatar; 
     if(!errors.isEmpty()){
-        fs.unlinkSync(`public/posts/${req.files[0].filename}`)
-        return res.json(errors.mapped())
+        if(postImg != postById.avatar){
+            fs.unlinkSync(`public/posts/${req.files[0].filename}`)
+        }
+        
+        res.json(errors.mapped())
     }else{
         try {
             await postModel.update({
